+++ gallery_dl/postprocessor/ugoira.py
@@ -1,6 +1,6 @@
 # -*- coding: utf-8 -*-
 
-# Copyright 2018-2021 Mike Fährmann
+# Copyright 2018-2022 Mike Fährmann
 #
 # This program is free software; you can redistribute it and/or modify
 # it under the terms of the GNU General Public License version 2 as
@@ -10,11 +10,19 @@
 
 from .common import PostProcessor
 from .. import util
-import collections
 import subprocess
 import tempfile
 import zipfile
+import shutil
 import os
+
+try:
+    from math import gcd
+except ImportError:
+    def gcd(a, b):
+        while b:
+            a, b = b, a % b
+        return a
 
 
 class UgoiraPP(PostProcessor):
@@ -27,18 +35,36 @@ class UgoiraPP(PostProcessor):
         self.output = options.get("ffmpeg-output", True)
         self.delete = not options.get("keep-files", False)
         self.repeat = options.get("repeat-last-frame", True)
+        self.mtime = options.get("mtime")
 
         ffmpeg = options.get("ffmpeg-location")
         self.ffmpeg = util.expand_path(ffmpeg) if ffmpeg else "ffmpeg"
 
+        mkvmerge = options.get("mkvmerge-location")
+        self.mkvmerge = util.expand_path(mkvmerge) if mkvmerge else "mkvmerge"
+
+        demuxer = options.get("ffmpeg-demuxer")
+        if demuxer is None or demuxer == "auto":
+            if self.extension in ("webm", "mkv") and (
+                    mkvmerge or shutil.which("mkvmerge")):
+                demuxer = "mkvmerge"
+            else:
+                demuxer = "concat" if util.WINDOWS else "image2"
+
+        if demuxer == "mkvmerge":
+            self._process = self._process_mkvmerge
+            self._finalize = self._finalize_mkvmerge
+        elif demuxer == "image2":
+            self._process = self._process_image2
+            self._finalize = None
+        else:
+            self._process = self._process_concat
+            self._finalize = None
+        self.log.debug("using %s demuxer", demuxer)
+
         rate = options.get("framerate", "auto")
         if rate != "auto":
             self.calculate_framerate = lambda _: (None, rate)
-
-        if options.get("ffmpeg-demuxer") == "image2":
-            self._process = self._image2
-        else:
-            self._process = self._concat
 
         if options.get("libx264-prevent-odd", True):
             # get last video-codec argument
@@ -88,13 +114,12 @@ class UgoiraPP(PostProcessor):
                 return
 
             # process frames and collect command-line arguments
-            args = self._process(tempdir)
+            pathfmt.set_extension(self.extension)
+            args = self._process(pathfmt, tempdir)
             if self.args:
                 args += self.args
-            self.log.debug("ffmpeg args: %s", args)
 
             # invoke ffmpeg
-            pathfmt.set_extension(self.extension)
             try:
                 if self.twopass:
                     if "-f" not in self.args:
@@ -105,48 +130,61 @@ class UgoiraPP(PostProcessor):
                 else:
                     args.append(pathfmt.realpath)
                     self._exec(args)
+                if self._finalize:
+                    self._finalize(pathfmt, tempdir)
             except OSError as exc:
                 print()
                 self.log.error("Unable to invoke FFmpeg (%s: %s)",
                                exc.__class__.__name__, exc)
                 pathfmt.realpath = pathfmt.temppath
             else:
+                if self.mtime:
+                    mtime = pathfmt.kwdict.get("_mtime")
+                    if mtime:
+                        util.set_mtime(pathfmt.realpath, mtime)
                 if self.delete:
                     pathfmt.delete = True
                 else:
                     pathfmt.set_extension("zip")
 
-    def _concat(self, path):
-        ffconcat = path + "/ffconcat.txt"
+    def _exec(self, args):
+        self.log.debug(args)
+        out = None if self.output else subprocess.DEVNULL
+        return subprocess.Popen(args, stdout=out, stderr=out).wait()
 
-        content = ["ffconcat version 1.0"]
-        append = content.append
-        for frame in self._frames:
-            append("file '{}'\nduration {}".format(
-                frame["file"], frame["delay"] / 1000))
-        if self.repeat:
-            append("file '{}'".format(frame["file"]))
-        append("")
-
-        with open(ffconcat, "w") as file:
-            file.write("\n".join(content))
-
+    def _process_concat(self, pathfmt, tempdir):
         rate_in, rate_out = self.calculate_framerate(self._frames)
         args = [self.ffmpeg, "-f", "concat"]
         if rate_in:
             args += ("-r", str(rate_in))
-        args += ("-i", ffconcat)
+        args += ("-i", self._write_ffmpeg_concat(tempdir))
         if rate_out:
             args += ("-r", str(rate_out))
         return args
 
-    def _image2(self, path):
-        path += "/"
+    def _process_image2(self, pathfmt, tempdir):
+        tempdir += "/"
+        frames = self._frames
+
+        # add extra frame if necessary
+        if self.repeat and not self._delay_is_uniform(frames):
+            last = frames[-1]
+            delay_gcd = self._delay_gcd(frames)
+            if last["delay"] - delay_gcd > 0:
+                last["delay"] -= delay_gcd
+
+                self.log.debug("non-uniform delays; inserting extra frame")
+                last_copy = last.copy()
+                frames.append(last_copy)
+                name, _, ext = last_copy["file"].rpartition(".")
+                last_copy["file"] = "{:>06}.{}".format(int(name)+1, ext)
+                shutil.copyfile(tempdir + last["file"],
+                                tempdir + last_copy["file"])
 
         # adjust frame mtime values
         ts = 0
-        for frame in self._frames:
-            os.utime(path + frame["file"], ns=(ts, ts))
+        for frame in frames:
+            os.utime(tempdir + frame["file"], ns=(ts, ts))
             ts += frame["delay"] * 1000000
 
         return [
@@ -155,18 +193,90 @@ class UgoiraPP(PostProcessor):
             "-ts_from_file", "2",
             "-pattern_type", "sequence",
             "-i", "{}%06d.{}".format(
-                path.replace("%", "%%"), frame["file"].rpartition(".")[2]),
+                tempdir.replace("%", "%%"),
+                frame["file"].rpartition(".")[2]
+            ),
         ]
 
-    def _exec(self, args):
-        out = None if self.output else subprocess.DEVNULL
-        return subprocess.Popen(args, stdout=out, stderr=out).wait()
+    def _process_mkvmerge(self, pathfmt, tempdir):
+        self._realpath = pathfmt.realpath
+        pathfmt.realpath = tempdir + "/temp." + self.extension
+
+        return [
+            self.ffmpeg,
+            "-f", "image2",
+            "-pattern_type", "sequence",
+            "-i", "{}/%06d.{}".format(
+                tempdir.replace("%", "%%"),
+                self._frames[0]["file"].rpartition(".")[2]
+            ),
+        ]
+
+    def _finalize_mkvmerge(self, pathfmt, tempdir):
+        args = [
+            self.mkvmerge,
+            "-o", self._realpath,
+            "--timecodes", "0:" + self._write_mkvmerge_timecodes(tempdir),
+        ]
+        if self.extension == "webm":
+            args.append("--webm")
+        args += ("=", pathfmt.realpath)
+
+        pathfmt.realpath = self._realpath
+        self._exec(args)
+
+    def _write_ffmpeg_concat(self, tempdir):
+        content = ["ffconcat version 1.0"]
+        append = content.append
+
+        for frame in self._frames:
+            append("file '{}'\nduration {}".format(
+                frame["file"], frame["delay"] / 1000))
+        if self.repeat:
+            append("file '{}'".format(frame["file"]))
+        append("")
+
+        ffconcat = tempdir + "/ffconcat.txt"
+        with open(ffconcat, "w") as file:
+            file.write("\n".join(content))
+        return ffconcat
+
+    def _write_mkvmerge_timecodes(self, tempdir):
+        content = ["# timecode format v2"]
+        append = content.append
+
+        delay_sum = 0
+        for frame in self._frames:
+            append(str(delay_sum))
+            delay_sum += frame["delay"]
+        append(str(delay_sum))
+        append("")
+
+        timecodes = tempdir + "/timecodes.tc"
+        with open(timecodes, "w") as file:
+            file.write("\n".join(content))
+        return timecodes
+
+    def calculate_framerate(self, frames):
+        uniform = self._delay_is_uniform(frames)
+        if uniform:
+            return ("1000/{}".format(frames[0]["delay"]), None)
+        return (None, "1000/{}".format(self._delay_gcd(frames)))
 
     @staticmethod
-    def calculate_framerate(framelist):
-        counter = collections.Counter(frame["delay"] for frame in framelist)
-        fps = "1000/{}".format(min(counter))
-        return (fps, None) if len(counter) == 1 else (None, fps)
+    def _delay_gcd(frames):
+        result = frames[0]["delay"]
+        for f in frames:
+            result = gcd(result, f["delay"])
+        return result
+
+    @staticmethod
+    def _delay_is_uniform(frames):
+        delay = frames[0]["delay"]
+        for f in frames:
+            if f["delay"] != delay:
+                return False
+        return True
 
 
 __postprocessor__ = UgoiraPP