+++ scripts/supportedsites.py
@@ -26,6 +26,7 @@ CATEGORY_MAP = {
     "bbc"            : "BBC",
     "bcy"            : "半次元",
     "comicvine"      : "Comic Vine",
+    "coomerparty"    : "Coomer",
     "deviantart"     : "DeviantArt",
     "drawfriends"    : "Draw Friends",
     "dynastyscans"   : "Dynasty Reader",
@@ -57,7 +58,9 @@ CATEGORY_MAP = {
     "imgur"          : "imgur",
     "joyreactor"     : "JoyReactor",
     "kabeuchi"       : "かべうち",
+    "kemonoparty"    : "Kemono",
     "kireicake"      : "Kirei Cake",
+    "kissgoddess"    : "Kiss Goddess",
     "lineblog"       : "LINE BLOG",
     "livedoor"       : "livedoor Blog",
     "omgmiamiswimwear": "Omg Miami Swimwear",
@@ -69,6 +72,7 @@ CATEGORY_MAP = {
     "mangapark"      : "MangaPark",
     "mangasee"       : "MangaSee",
     "mastodon.social": "mastodon.social",
+    "mememuseum"     : "meme.museum",
     "myhentaigallery": "My Hentai Gallery",
     "myportfolio"    : "Adobe Portfolio",
     "naverwebtoon"   : "NaverWebtoon",
@@ -121,12 +125,15 @@ CATEGORY_MAP = {
 }
 
 SUBCATEGORY_MAP = {
+    "art"    : "Art",
+    "audio"  : "Audio",
     "doujin" : "Doujin",
     "gallery": "Galleries",
     "image"  : "individual Images",
     "index"  : "Site Index",
     "issue"  : "Comic Issues",
     "manga"  : "Manga",
+    "media"  : "Media Files",
     "popular": "Popular Images",
     "recent" : "Recent Images",
     "search" : "Search Results",
@@ -143,6 +150,10 @@ SUBCATEGORY_MAP = {
     },
     "atfbooru": {
         "favorite": "",
+    },
+    "coomerparty": {
+        "discord"       : "",
+        "discord-server": "",
     },
     "danbooru": {
         "favorite": "",
@@ -172,14 +183,10 @@ SUBCATEGORY_MAP = {
     "mangadex": {
         "feed" : "Followed Feed",
     },
-    "newgrounds": {
-        "art"  : "Art",
-        "audio": "Audio",
-        "media": "Media Files",
-    },
     "pinterest": {
         "board": "",
         "pinit": "pin.it Links",
+        "created": "Created Pins",
     },
     "pixiv": {
         "me"  : "pixiv.me Links",
@@ -238,6 +245,7 @@ AUTH_MAP = {
     "aryion"         : "Supported",
     "atfbooru"       : "Supported",
     "baraag"         : _OAUTH,
+    "coomerparty"    : "Supported",
     "danbooru"       : "Supported",
     "derpibooru"     : _APIKEY_DB,
     "deviantart"     : _OAUTH,
@@ -266,7 +274,7 @@ AUTH_MAP = {
     "ponybooru"      : "API Key",
     "reddit"         : _OAUTH,
     "sankaku"        : "Supported",
-    "seiga"          : "Required",
+    "seiga"          : _COOKIES,
     "seisoparty"     : "Supported",
     "smugmug"        : _OAUTH,
     "subscribestar"  : "Supported",
@@ -379,6 +387,10 @@ def build_extractor_list():
     # add e-hentai.org
     default["e-hentai"] = default["exhentai"]
     domains["e-hentai"] = domains["exhentai"].replace("x", "-")
+
+    # add coomer.party
+    default["coomerparty"] = default["kemonoparty"]
+    domains["coomerparty"] = domains["kemonoparty"].replace("kemono", "coomer")
 
     # add hentai-cosplays sister sites (hentai-img, porn-images-xxx)
     default["hentaiimg"] = default["hentaicosplays"]